Improve setting of column default value
Split into 2 methods for setting either a quoted or raw value for the default. Use SqlFragment to indicate that the default is a raw value (rather than storing the `isConstant` flag alongside the value.

Remove ability to pass `null` for default to "unset" the default value - full column definitions must always be provided anyway, so just omitting the call to set a default will remove the column default in an alter.