Change binary path argument to a string
This was originally an array argument due to a convention from a previous project of always using an array for binary paths when using Symfony process, so that it is possible to specify an interpreter/script combination.

This doesn't really make sense for the online schema change binary, and just adds confusion for the public API, so better to simplify it to a string.